# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "Endpoint to go to homepage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users": {
      "get": {
        "description": "Endpoint to get all users.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Endpoint to get the specific user.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/edit/{id}": {
      "put": {
        "description": "Endpoint to edit the specific user.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nom": {
                  "example": "Anthony KHELIL"
                },
                "adresse": {
                  "example": "2 rue du cimetière"
                },
                "email": {
                  "example": "anthony.khelil@ynov.com"
                },
                "pays": {
                  "example": "France"
                },
                "status": {
                  "example": "Inactif"
                },
                "role": {
                  "example": "Client"
                },
                "id": {
                  "example": "12345678"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/delete": {
      "delete": {
        "description": "Endpoint to delete the specific user.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "123456789"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/add": {
      "post": {
        "description": "Endpoint to add user.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nom": {
                  "example": "Saleh DASSOUKHI"
                },
                "adresse": {
                  "example": "25 Rue d'adresse"
                },
                "email": {
                  "example": "email@email.fr"
                },
                "pays": {
                  "example": "France"
                },
                "status": {
                  "example": "Actif"
                },
                "role": {
                  "example": "Admin"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}